chore: add deriv x modal
